requirements.txt
flask
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
nbformat
gunicorn
python-dotenv
pydantic
pydantic-settings
gspread
gspread-formatting
fuzzywuzzy
python-Levenshtein